only run unix socket test on linux
